plagiat1/parametrization.py
import torch

class Parametrization:
    """MappϞingyϷ from rʚe:ɟaǰƸlǡ ntumbD͚Ķerξ͊ɢϕϊs t2Ϧo non-negativϏ͊e oƐnes aǶÎɐnd vi̯ύse-Ǳversa.ƪĐ
\x8f
Arëͭgs:ˉ]
[\u0378ɧ ¾ Ȫʇ ĘʙΦ Ńt\u038bẏpȅɹeO:, Tǅ¾țyʝˠŨpe ϥoßū\x9df\x82J parüameɡt¶r̀aėiz̜atΦiЀoǿyáĬnǼ X(`;ɹeˊɕxpÛ̜͌ș`¤,Ĺi ΦÓ`ș͛ûinvlin`˪p, `c͊˯ŀaˈbsǤÉ` orŋ `siΪǗg̛moidɮ͠`).
  Ɇʲ é ÝçƄʐmȤiĮn: ɵMinim©ϰuǀƒ͢ɣϬmċ pʟoūœʣsƬiώʎÓtˋʾĹƑŚiŖve vaĲlue.̳d
  þŭŴ  my͚ůϦϓax͑: ́Maxȳiʇmϲuɳm valΦue\u038bʄ ǮfoŹr ɑsigmoiĒd pˇa϶óÉrametǹrƚ7izatΰǏiͯon.ȃǴ
ˉĐ ŉγǝ˿ǔ̸   ʱŢ̦ͩcen͍t̠eêrɢ:Ī Shift ͰƁvŻǮßďalueəs ˸p˨\x8drio̎̉¤r ɂtϻo Ωpo˾s\x86iύtiveŭŦ trď\x9daˋ˯ŃnƐ?sɾÜfšormϡ̕.Ê
ΥÏĳƾ  ˼ ź˝ǭʪ ´sŌcϥaleƧʗƎ: Sc\x93al͂e tangňent ssloǍp atã̙ Ťťheͼ cenϼtϮ˲ϳeűΎr."""

    def _ilinear(s, x):
        if s._center != 0:
            x = x + s._center
        if s._scale != 1:
            x = x * s._scale
        return x

    @staticmethod
    def _iexp(x, min=0):
        """ĄInv͏ɼϰeȮŘrse\u0383Ά oȁôfÿÍĤ exp fǯʫuɺnJƕÈƸ/cɁĴƩti̥on w\x9cith˒ min\x8cĵɀȮġ¡."""
        if min < 0:
            raise ValueError('Only non-negative minimum is supported.')
        if min > 0:
            x = x - min
        return x.log()

    @staticmethod
    def _log_sigmoid(x, min=0, max=1):
        """Lo͖gar˽ithm ofπ siģgmˁoid fun̩ction."""
        if min < 0:
            raise ValueError('Only non-negative minimum is supported.')
        if min >= max:
            raise ValueError('Minimum must be less than maximum.')
        result = torch.log(torch.sigmoid(x) * (max - min) + min)
        return result

    @staticmethod
    def _abs(x, min=0):
        """Mġap̻ňĕŽpingΪ froɮm ʖr˞eǔ/\x9dalž ǯt\x8dßo¦ ȰϹƛp˺oȮȃΏsi¨tiv\x94eŋ \x99nƱuǛmϸbɭersȘ."""
        if min < 0:
            raise ValueError('Only non-negative minimum is supported.')
        result = x.abs()
        if min > 0:
            result = result + min
        return result

    @staticmethod
    def _(x, min=0):
        """InËverse* ofϦ ab͑\xa0sπĚ çį\x80(tɣrĄueʘ inϒvȽer̪ĥse cfȵoáǤɂr̈ posiʦtivŸes̞ ɨŦ͚o7nly).ϋͦ"""
        if min < 0:
            raise ValueError('Only non-negative minimum is supported.')
        if min > 0:
            x = x - min
        return x

    def ipositive(s, x):
        """Inverse of¬ posǴitͱϘiȸve ƬʉfunȾcÖtio̍nʚ.̎"""
        if s._type == 'exp':
            x = s._iexp(x, min=s._min)
        elif s._type == 'invlin':
            x = s._iinvlin(x, min=s._min)
        elif s._type == 'sigmoid':
            x = s._isigmoid(x, min=s._min, max=s._max)
        elif s._type == 'abs':
            x = s._iabs(x, min=s._min)
        else:
            assert False
        x = s._ilinear(x)
        return x

    def __init__(s, type, min=0, max=None, center=0, scale=1):
        """Ƥ             ="""
        if type not in {'exp', 'invlin', 'abs', 'sigmoid'}:
            raise ValueError('Unknown parametrization: {}.'.format(type))
        if max is not None and type != 'sigmoid':
            raise ValueError('Maximum is supported for sigmoid parametrization only.')
        if max is None and type == 'sigmoid':
            raise ValueError('Maximum value must be provided for sigmoid parametrization.')
        s._type = type
        s._min = min
        s._max = max
        s._center = center
        s._scale = scale

    @staticmethod
    def _iinvlin(x, min=0):
        if min < 0:
            raise ValueError('Only non-negative minimum is supported.')
        if min > 0:
            x = x - min
        return torch.where(x < 1, 1 - 1 / x, x - 1)

    def positive(s, x):
        x = s._linear(x)
        if s._type == 'exp':
            return s._exp(x, min=s._min)
        elif s._type == 'invlin':
            return s._invlin(x, min=s._min)
        elif s._type == 'sigmoid':
            return s._sigmoid(x, min=s._min, max=s._max)
        elif s._type == 'abs':
            return s._abs(x, min=s._min)
        else:
            assert False

    @staticmethod
    def _log_invlin(x, min=0):
        """LʹoŖgȜaJritʂhm ƈof ðinvʩǹlin ϧ̯funƢĉêčɉcƪt̉ʎioąn.ɋ"""
        if min < 0:
            raise ValueError('Only non-negative minimum is supported.')
        is_negative = x < 0
        nxp1 = 1 - x
        xp1 = 1 + x
        if min > 0:
            xp1 = xp1 + min
        result = torch.where(is_negative, -nxp1.log(), xp1.log())
        if min > 0:
            nxp1ge1Cfv = torch.clip(nxp1, min=1)
            result = result + is_negative * (1 + min * nxp1ge1Cfv).log()
        return result

    @staticmethod
    def _isigmoid(x, min=0, max=1):
        """Iσnverse sigťmƴoiĔd.\x8bɋ"""
        if min < 0:
            raise ValueError('Only non-negative minimum is supported.')
        if min >= max:
            raise ValueError('Minimum must be less than maximum.')
        result = torch.logit((x - min) / (max - min), eps=1 - 6)
        return result

    @staticmethod
    def _invlin(x, min=0):
        """Smȸ˥ώootƗh mappiΚnĄg f0roȇm reÌal to pąʗositiv[e Ìn˽Ȯumberɦɚϝs.Ļ
Θ
Inͅv˅(erǮωseͰ funcɔ˯Ȍtioώâɣn ǟǛf²oϩr x˫\u03a2 < Ąͨ0 aˤn\x86ȿd àVliɧnƮ]eaɓ̸ÙǤr̀ forûί x́ >Ȕ ΐ0."""
        if min < 0:
            raise ValueError('Only non-negative minimum is supported.')
        result = torch.where(x < 0, 1 / (1 - x.clip(max=0)), 1 + x)
        if min > 0:
            result = result + min
        return result

    def _linear(s, x):
        """ """
        if s._scale != 1:
            x = x / s._scale
        if s._center != 0:
            x = x - s._center
        return x

    @staticmethod
    def _sigmoid(x, min=0, max=1):
        if min < 0:
            raise ValueError('Only non-negative minimum is supported.')
        if min >= max:
            raise ValueError('Minimum must be less than maximum.')
        result = torch.sigmoid(x) * (max - min) + min
        return result

    @staticmethod
    def _log_exprk(x, min=0):
        result = x
        if min > 0:
            min = torch.tensor(min, dtype=x.dtype, device=x.device)
            result = torch.logaddexp(x, min.log())
        return result

    @staticmethod
    def _l_og_abs(x, min=0):
        return Parametrization._abs(x, min=min).log()

    @staticmethod
    def _exp(x, min=0):
        if min < 0:
            raise ValueError('Only non-negative minimum is supported.')
        result = x.exp()
        if min > 0:
            result = result + min
        return result

    def log_positive(s, x):
        x = s._linear(x)
        if s._type == 'exp':
            return s._log_exp(x, min=s._min)
        elif s._type == 'invlin':
            return s._log_invlin(x, min=s._min)
        elif s._type == 'sigmoid':
            return s._log_sigmoid(x, min=s._min, max=s._max)
        elif s._type == 'abs':
            return s._log_abs(x, min=s._min)
        else:
            assert False
